Simulate counts from a negative binomial model

Given an object that describes negative binomial distribution, simulate counts from the distribution.

newSim(object, seed, ...)

# S4 method for newmodel
newSim(object, seed)

Arguments

object
an object that describes a matrix of negative binomial.

seed
an optional integer to specify how the random number generator should be initialized with a call to set.seed. If missing, the random generator state is not changed.

...
additional arguments.

Value

A list with the following elements.

countsthe matrix with the simulated counts.

dataNBthe data simulated from the negative binomial.

dataDropoutsthe data simulated from the binomial process.

zeroFractionthe fraction of zeros.

Methods (by class)

newmodel: simulate from a nb distribution.

Examples

a <- newmodel(n=5, J=10)
newSim(a)

#> $counts
#>       [,1] [,2] [,3] [,4] [,5]
#>  [1,]    1    0    0    0    2
#>  [2,]    0    1    3    0    2
#>  [3,]    6    1    0    0    1
#>  [4,]    1    1    0    1    0
#>  [5,]    0    0    2    0    2
#>  [6,]    1    2    0    0    1
#>  [7,]    2    3    0    1    0
#>  [8,]    0    0    3    1    2
#>  [9,]    0    0    2    0    0
#> [10,]    1    0    1    0    1
#> 
#> $zeroFraction
#> [1] 0.48
#> 
#> attr(,"seed")
#>   [1]       10403         581 -2089519707 -1093835428 -1319683790  1392429101
#>   [7]  -202324249   631843854  -539511320  1353072619 -1303625591   879318040
#>  [13]   774136646  1478877601 -1644826573 -1155924574 -1612494700 -1731969257
#>  [19]  -845694707   474058788 -1674022902 -1410673291   -72657713   325212118
#>  [25]  1081710464 -1118188605 -1231921823   521389008  1462334254   265391833
#>  [31]  1689205995 -1189593766  -976466148   404318751  -837617035 -1825212660
#>  [37]  1573186274      315325 -1025460169 -1853639970  2095787544  -982486661
#>  [43]   922061081  1343641000   946764118   530950609   329083267  -334890990
#>  [49]  -309281884  1177791335  1832358077  1128990772  1553173402    33392613
#>  [55]  1738486655 -1895238906   316341392 -1461736141 -1293989167 -1084706688
#>  [61]   251251038  -199237975  -361027237   976855914 -1401058804   295033423
#>  [67]  1231140613   528547772   703052434  -305520243   510577095  -706028882
#>  [73]  2077389576  1635348747    55128041 -1622058824  2080255846  1787411457
#>  [79]   -51320173   802553154 -1422732556 -1690329353  1321746157 -1152358012
#>  [85]  -764057686 -1690460907    -2386001  1579456438   736030176 -1761295069
#>  [91]  1204959809 -1338288848   517082958  1004819897    84322251  1727118458
#>  [97]  1730305404  -606472769   214872725   171765932   317487746   736686685
#> [103]  -746781353 -1892652994 -1246229384   757118811   935006905   955395848
#> [109]   584918454 -1957668111  1740918051 -1430395022  2039604676   720149895
#> [115]   651857757  1196065108 -1502727302 -1034624315  -266779617    50901926
#> [121] -1763801808 -2141575213 -1923070479  1952256800  -113943106 -1467755063
#> [127]   352987515   552218442  2003816364  -193538385  1253409893  2102928412
#> [133]   562907762  -571183891   -86682329  1801396558  1474711976  1427129771
#> [139]  -293388599  -736147752   147016582  2111946337 -1722873869  1328929378
#> [145] -1539654444  1074716119  -418916403 -1213562012   504661834  -103678283
#> [151]  2024568335   449611542   860488128  -624893565  2123659169 -1054541936
#> [157] -1105572754  2081544729  -574078293   310023834 -1294876068    80445279
#> [163]  1273172277   -82038580  1862769186 -1398549379 -1387106953  -990461282
#> [169] -1426648744 -1713829701 -1954179495   615106024   155041814   254585745
#> [175]  -705162173  -831411374   187232228   471871015  -888625539   507475828
#> [181]  -294327462  2056830373  -185566785   737161798  -378527536 -1976665357
#> [187]  -860881903  -695268800  2052148766  -660961047  -789525989 -1304724950
#> [193]  -804098740  1535520527  -319753659 -1779084292   569674322 -1049561139
#> [199]   161472903 -2058800018 -1525257144  1755321931 -1768739927  -715414792
#> [205]     7357990   423108417 -1675451181  1063069826  -750205260  1278960183
#> [211] -1847301843   129108548  1840789354 -1792728107 -1537919633  -884576138
#> [217]  1817697184  -661308317  -721073919  -601373584  1223450638  -895926151
#> [223]  -143884277  1150554426  1049858108 -1911449729  1664396501  -798977812
#> [229]  -881391038 -2069615046   366650576 -1676327828  -804498764  -162584558
#> [235]  -194488352  -875753524  -414159392 -2076660190  -687003352   866251020
#> [241]    -8341956  -256450318   358513392 -1741921468   807610968 -2139631046
#> [247] -1474853136  -713196356   138677780 -1378869054 -1280872352  1371901308
#> [253]  2142051536 -1801233374 -1719002456    73396108  -297231940  1028486194
#> [259]   550557504   -56198988 -1751399608  -411559238   854740752   220272876
#> [265]   -25693100 -1378482414   490468256   812851148  1581476704   181252290
#> [271] -2066770392   513459628 -1708974276  1235447666   100107056 -1536556508
#> [277] -1383526344  1134472410  -339861104  1590241020 -1306247276  1209819970
#> [283]  1366759936  1898349116   328939088 -1305783646   662203208  1797044748
#> [289]  1610476764 -1687362990  -212395904 -1224492204  -713322008   559237882
#> [295]  -405006512   954139692   505214324   490499730  1340488288  1229255308
#> [301] -1863537824  1818702562   975714472  1148195020 -1370516996  -983433294
#> [307]   -88287760  -597635644   364857176  1640995706   273128880  -335494724
#> [313]  1636708436 -2060594238 -1170771168   229463996   728928848   668859874
#> [319]   677036392  1111928140  -206753540  2049819378  -790233344  1091985012
#> [325]  2080673800   246677946 -1178755120  1653250540 -1379993708 -1335713070
#> [331]  1969700448 -1397119668 -1324866272  1754892930 -1454147096  1411596268
#> [337]   426985020  1913419378  1679739760  1909504292  -116747720   110963098
#> [343]   -35808816 -1397766724   922548372  -319220606  -833256000 -1311552708
#> [349]   516033680  -818918622 -1795541496  -888834804   987284892  1688419858
#> [355] -1948021888   462971668   799850792  1446872122  -451114032  1729178220
#> [361]   415539764 -1887638894  -452450976   522451276  -190371872  -278751838
#> [367]  1875002792   976422540 -1350025796 -1507819918  -120736784  -821113276
#> [373] -1402511272 -1632442310  1157165168  1417615420  1315234068   132186818
#> [379]  -687487264  1526511740  1807029968   773101986   695990312  -914470132
#> [385] -1815967300  -506158286  -144946496 -1947931596  -325402936   189527226
#> [391]   576951952   755150572  1661786708  -894532974  1348585760  -365325876
#> [397]   581550816 -1196244798 -1796191192   640489388  -183555396  -655696910
#> [403]  1798592048   264212388 -1200370248  -745220262 -1625599856  -655733892
#> [409]  -924966636 -1497773758 -1593569024 -1054188100  1969627472  -708732638
#> [415]   305797704  1754099724  1174709852  -924710446  1442002432  -632625836
#> [421]  -848778008 -1277587206  2082154192   121236268  2039492852  1689966994
#> [427]   519732576  -752985076 -1360085792 -1568886174  1680565416  1031535692
#> [433]  -176879748  -954427086 -1853267344   -39918780 -1838529576  -777170054
#> [439]  -229926864   859754428   161722964  -273403838  1743761056  1051902652
#> [445]   491391952  -991871646  1729253096   162049228  1554915836 -1300611598
#> [451]  -448240512  1801333236  1995295368   761387834     2035024 -1000870804
#> [457]  1922709043  1811168405  1145930994 -1017977536  1700997801  -625073381
#> [463]  1214436588 -1545824006   526158127 -1687098055  -854417586 -2049149540
#> [469]   476681997  1331290487   396408192  1467188190  1690758315 -1073154131
#> [475]  1731808282   117867224   317033281 -1492620301 -1728072172   719630242
#> [481]  1882755255 -1858155903  -355296634  -381134140 -1924899851  1789522527
#> [487]   178947256   -55574122   868154275  1758945125 -1244612990 -2016812048
#> [493]   147781017  1938561931 -1052658052   486796458 -1163985185  -846007191
#> [499] -1483239810 -1212930452 -1174787395   819206023   228643696  -508648946
#> [505]  -392532165 -2102471651 -1286054710  -881137688 -1954974511   871186627
#> [511]  -824727612  1678623858   191790407   151032081 -1476660074  1621117428
#> [517]   653007749  -961056049  1516318664  -731748762 -1833531245  -506939403
#> [523]  -360336942  1578023456 -2010115063   666487099 -1212138356  -599860966
#> [529]  -744464497  2051088601 -1615859858 -1430382852  1546474605 -1135218665
#> [535]  1089432032   193418686 -1509557365  1183190029  1781349434  1118555256
#> [541]   322979873 -1740291373 -2025082380   229251714  -570460393  1874197985
#> [547] -1503562970 -1510117852  -249456427 -2009941889 -2103812840   286463030
#> [553]  -965297469 -1263216891  -898299230 -1627805552   664033337  1090365675
#> [559]   692816348   -68775926 -1542123777  -813719735 -1923066658   732228236
#> [565] -1137770915  1320846439 -1535735408   -39364178 -2121450405 -2099192771
#> [571]   858550954   578752328   753795953 -1353243421   448211940   695761682
#> [577]  1507702503 -1202085711  1986908918   597439764  -782759387 -1437976785
#> [583]  -759224088   251439558 -1533859725 -1525447979 -1499602766   244180096
#> [589]  1099733097  -214174501 -1127532500  -907646918   408889071   452364665
#> [595]  -180087282  -264398884  2007531597  -647191625  2048056640 -1865171938
#> [601]  1878239979   845734509  -653176102  1157638808 -1754224383  -971870541
#> [607]  -937096748  2136844514 -1188630153  1398403265    15404102 -1738157180
#> [613]   369886517 -1811971297  1621862008  -915793066  -302456349  1414690725
#> [619]   -73656766  1091099696  1683357785  1066823115   313056828  1541374058
#> [625]  1713650335   205702693