Class newmodel

Objects of this class store all the values needed to work with a negative binomial model, as described in the vignette. They contain all information to fit a model by penalized maximum likelihood or simulate data from a model.

# S4 method for newmodel
show(object)

# S4 method for newmodel
numberSamples(x)

# S4 method for newmodel
numberFeatures(x)

# S4 method for newmodel
numberFactors(x)

# S4 method for newmodel
newX(object)

# S4 method for newmodel
newV(object)

# S4 method for newmodel
newLogMu(object)

# S4 method for newmodel
newMu(object)

# S4 method for newmodel
newZeta(object)

# S4 method for newmodel
newPhi(object)

# S4 method for newmodel
newTheta(object)

# S4 method for newmodel
newEpsilon_beta(object)

# S4 method for newmodel
newEpsilon_gamma(object)

# S4 method for newmodel
newEpsilon_W(object)

# S4 method for newmodel
newEpsilon_alpha(object)

# S4 method for newmodel
newEpsilon_zeta(object)

# S4 method for newmodel
newW(object)

# S4 method for newmodel
newBeta(object)

# S4 method for newmodel
newGamma(object)

# S4 method for newmodel
newAlpha(object)

Arguments

object
an object of class newmodel.

x
an object of class newmodel.

Value

numberSamples returns the number of samples; numberFeaturesreturns the number of features; numberFactors returns the number of latent factors.

Details

For the full description of the model see the model vignette. Internally, the slots are checked so that the matrices are of the appropriate dimensions: in particular, X, O and W need to have n rows, V needs to have J rows, zeta must be of length J.

Methods (by generic)

show: show useful info on the object.

numberSamples: returns the number of samples.

numberFeatures: returns the number of features.

numberFactors: returns the number of latent factors.

newX: returns the sample-level design matrix for mu.

newV: returns the gene-level design matrix for mu.

newLogMu: returns the logarithm of the mean of the non-zero component.

newMu: returns the mean of the non-zero component.

newZeta: returns the log of the inverse of the dispersion parameter.

newPhi: returns the dispersion parameter.

newTheta: returns the inverse of the dispersion parameter.

newEpsilon_beta: returns the regularization parameters for beta.

newEpsilon_gamma: returns the regularization parameters for gamma.

newEpsilon_W: returns the regularization parameters for W.

newEpsilon_alpha: returns the regularization parameters for alpha.

newEpsilon_zeta: returns the regularization parameters for zeta.

newW: returns the matrix W of inferred sample-level covariates.

newBeta: returns the matrix beta of inferred parameters.

newGamma: returns the matrix gamma of inferred parameters.

newAlpha: returns the matrix alpha of inferred parameters.

Slots

X

matrix. The design matrix containing sample-level covariates, one sample per row.

V

matrix. The design matrix containing gene-level covariates, one gene per row.

X_intercept

logical. TRUE if X contains an intercept.

V_intercept

logical. TRUE if V contains an intercept.

W

matrix. The factors of sample-level latent factors.

beta

matrix or NULL. The coefficients of X in the regression.

gamma

matrix or NULL. The coefficients of V in the regression.

alpha

matrix. The weight of sample-level latent factors.

zeta

numeric. A vector of log of inverse dispersion parameters.

epsilon_beta

nonnegative scalar. Regularization parameter for beta

epsilon_gamma

nonnegative scalar. Regularization parameter for gamma

epsilon_W

nonnegative scalar. Regularization parameter for W

epsilon_alpha

nonnegative scalar. Regularization parameter for alpha

epsilon_zeta

nonnegative scalar. Regularization parameter for zeta