Optimize the parameters of a Negative Binomial regression model

The parameters of the model given as argument are optimized by penalized maximum likelihood on the count matrix given as argument.

optimization(
  Y,
  cluster,
  children,
  model,
  max_iter,
  stop_epsilon,
  n_gene_disp,
  n_cell_par,
  n_gene_par,
  commondispersion,
  verbose
)

Arguments

cluster
The PSOCK cluster

children
Number of child process

model
newmodel item

max_iter
maximum number of iterations

stop_epsilon
stopping criterion, when the relative gain in likelihood is below epsilon

n_gene_disp
number of genes used in mini-batch dispersion estimation approach(default NULL > all genes are used)

n_cell_par
number of cells used in mini-batch cell's related parameters estimation approach(default NULL > all cells are used)

n_gene_par
number of genes used in mini-batch gene's related parameters estimation approach(default NULL > all genes are used)

commondispersion
Whether or not a single dispersion for all features is estimated (default TRUE).

verbose
print information (default FALSE)

Value

An object of class newmodel similar to the one given as argument with modified parameters alpha, beta, gamma, W.